% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\docType{data}
\name{mxstate.map}
\alias{mxstate.map}
\title{Map of the 31 Mexican states plus the Federal District}
\usage{
data(mxstate.map)
}
\description{
A data.frame which contains a map of all 31 Mexican States plus
the Federal District.  The shapefile
was modified using QGIS in order to remove
the María islands, Socorro island, Arrecife Alacrán and
Guadalupe Island, then simplified with http://www.mapshaper.org/
}
\examples{
\dontrun{
# render the map with ggplot2
library(ggplot2)

data(mxstate.map)
ggplot(mxstate.map, aes(long, lat, group=group)) + geom_polygon(color = "black")
}
}
\references{
Downloaded from the "Cartografía Geoestadística Urbana y
Rural Amanzanada. Planeación de la Encuesta Intercensal 2015" shapefiles
(https://gist.github.com/diegovalle/aa3eef87c085d6ea034f)
}

